requirements.txt
asgiref==3.9.0
cachetools==5.5.2
dj-database-url==3.0.1
Django==5.2.4
django-debug-toolbar==5.2.0
django-environ==0.12.0
django-storages==1.14.6
djangorestframework==3.16.0
google-auth==2.40.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dotenv==1.1.1
rsa==4.9.1
sqlparse==0.5.3
django-redis==5.4.0
requests==2.32.3
pgvector==0.4.1
django-allauth==0.61.1
geopy==2.4.1
django-taggit>=4.0.0
